refactor: keep mobile adapters out of contract
